finally got local rendering to look right
