+++ src/components/hybrid/compositor/compositor-pass.js
@@ -1,4 +1,4 @@
-import {Pass} from './effect-composer';
+import {FullScreenQuad, Pass} from './pass';
 import {CompositorShader} from './compositor-shader';
 
 export class CompositorPass extends Pass {
@@ -30,31 +30,7 @@ export class CompositorPass extends Pass {
 
         this.needsSwap = false;
 
-        this.quadCamera = new THREE.OrthographicCamera(-1, 1, 1, -1, 0, 1);
-
-        /* this.quadCameraL = new THREE.OrthographicCamera(-1, 1, 1, -1, 0, 1);
-        this.quadCameraL.layers.enable( 1 );
-        this.quadCameraR = new THREE.OrthographicCamera(-1, 1, 1, -1, 0, 1);
-        this.quadCameraR.layers.enable( 2 );
-
-        this.quadCameraVR = new THREE.ArrayCamera();
-        this.quadCameraVR.layers.enable( 1 );
-        this.quadCameraVR.layers.enable( 2 );
-        this.quadCameraVR.cameras = [ this.quadCameraL, this.quadCameraR ]; */
-
-        this.fsQuad = new THREE.Mesh(new THREE.PlaneGeometry( 2, 2 ), this.material);
-        this.quadScene = new THREE.Scene();
-        this.quadScene.add(this.fsQuad);
-
-        /* this.fsQuadL = new THREE.Mesh(new THREE.PlaneGeometry( 2, 2 ), this.material);
-        this.fsQuadL.geometry.attributes.uv.array.set([0, 1, 0.5, 1, 0, 0, 0.5, 0]);
-        this.quadSceneL = new THREE.Scene();
-        this.quadSceneL.add(this.fsQuadL);
-
-        this.fsQuadR = new THREE.Mesh(new THREE.PlaneGeometry( 2, 2 ), this.material);
-        this.fsQuadR.geometry.attributes.uv.array.set([0.5, 1, 1, 1, 0.5, 0, 1, 0]);
-        this.quadSceneR = new THREE.Scene();
-        this.quadSceneR.add(this.fsQuadR); */
+        this.fsQuad = new FullScreenQuad(this.material);
 
         window.addEventListener('enter-vr', this.onEnterVR.bind(this));
         window.addEventListener('exit-vr', this.onExitVR.bind(this));
@@ -90,8 +66,7 @@ export class CompositorPass extends Pass {
         this.material.uniforms.tDiffuse.value = readBuffer.texture;
         this.material.uniforms.tDepth.value = readBuffer.depthTexture;
 
-        renderer.setRenderTarget(readBuffer);
-        renderer.render(this.scene, this.camera);
-        renderer.setRenderTarget(null);
+        renderer.setRenderTarget( writeBuffer );
+        this.fsQuad.render( renderer );
     }
 }
