--- src/components/hybrid/compositor/effect-composer.js
@@ -1,504 +1,0 @@
-/**
- * Full-screen textured quad shader
- */
-
-const CopyShader = {
-
-	uniforms: {
-
-		'tDiffuse': { value: null },
-		'opacity': { value: 1.0 }
-
-	},
-
-	vertexShader: /* glsl */`
-
-		varying vec2 vUv;
-
-		void main() {
-
-			vUv = uv;
-			gl_Position = projectionMatrix * modelViewMatrix * vec4( position, 1.0 );
-
-		}`,
-
-	fragmentShader: /* glsl */`
-
-		uniform float opacity;
-
-		uniform sampler2D tDiffuse;
-
-		varying vec2 vUv;
-
-		void main() {
-
-			gl_FragColor = texture2D( tDiffuse, vUv );
-			gl_FragColor.a *= opacity;
-
-
-		}`
-
-};
-
-// Helper for passes that need to fill the viewport with a single quad.
-
-const _camera = new THREE.OrthographicCamera( - 1, 1, 1, - 1, 0, 1 );
-
-// https://github.com/mrdoob/three.js/pull/21358
-
-const _geometry = new THREE.BufferGeometry();
-_geometry.setAttribute( 'position', new THREE.Float32BufferAttribute( [ - 1, 3, 0, - 1, - 1, 0, 3, - 1, 0 ], 3 ) );
-_geometry.setAttribute( 'uv', new THREE.Float32BufferAttribute( [ 0, 2, 0, 0, 2, 0 ], 2 ) );
-
-class FullScreenQuad {
-
-	constructor( material ) {
-
-		this._mesh = new THREE.Mesh( _geometry, material );
-
-	}
-
-	dispose() {
-
-		this._mesh.geometry.dispose();
-
-	}
-
-	render( renderer ) {
-
-		renderer.render( this._mesh, _camera );
-
-	}
-
-	get material() {
-
-		return this._mesh.material;
-
-	}
-
-	set material( value ) {
-
-		this._mesh.material = value;
-
-	}
-
-}
-
-class Pass {
-
-    constructor() {
-
-        // if set to true, the pass is processed by the composer
-        this.enabled = true;
-
-        // if set to true, the pass indicates to swap read and write buffer after rendering
-        this.needsSwap = true;
-
-        // if set to true, the pass clears its buffer before rendering
-        this.clear = false;
-
-        // if set to true, the result of the pass is rendered to screen. This is set automatically by EffectComposer.
-        this.renderToScreen = false;
-
-    }
-    setSize() {}
-    render() {
-
-        console.error( 'THREE.Pass: .render() must be implemented in derived pass.' );
-
-    }
-
-}
-
-class ShaderPass extends Pass {
-
-	constructor( shader, textureID ) {
-
-		super();
-
-		this.textureID = ( textureID !== undefined ) ? textureID : 'tDiffuse';
-
-		if ( shader instanceof THREE.ShaderMaterial ) {
-
-			this.uniforms = shader.uniforms;
-
-			this.material = shader;
-
-		} else if ( shader ) {
-
-			this.uniforms = THREE.UniformsUtils.clone( shader.uniforms );
-
-			this.material = new THREE.ShaderMaterial( {
-
-				defines: Object.assign( {}, shader.defines ),
-				uniforms: this.uniforms,
-				vertexShader: shader.vertexShader,
-				fragmentShader: shader.fragmentShader
-
-			} );
-
-		}
-
-		this.fsQuad = new FullScreenQuad( this.material );
-
-	}
-
-	render( renderer, writeBuffer, readBuffer /*, deltaTime, maskActive */ ) {
-
-		if ( this.uniforms[ this.textureID ] ) {
-
-			this.uniforms[ this.textureID ].value = readBuffer.texture;
-
-		}
-
-		this.fsQuad.material = this.material;
-
-		if ( this.renderToScreen ) {
-
-			renderer.setRenderTarget( null );
-			this.fsQuad.render( renderer );
-
-		} else {
-
-			renderer.setRenderTarget( writeBuffer );
-			// TODO: Avoid using autoClear properties, see https://github.com/mrdoob/three.js/pull/15571#issuecomment-465669600
-			if ( this.clear ) renderer.clear( renderer.autoClearColor, renderer.autoClearDepth, renderer.autoClearStencil );
-			this.fsQuad.render( renderer );
-
-		}
-
-	}
-
-	dispose() {
-
-		this.material.dispose();
-
-		this.fsQuad.dispose();
-
-	}
-
-}
-
-class MaskPass extends Pass {
-
-	constructor( scene, camera ) {
-
-		super();
-
-		this.scene = scene;
-		this.camera = camera;
-
-		this.clear = true;
-		this.needsSwap = false;
-
-		this.inverse = false;
-
-	}
-
-	render( renderer, writeBuffer, readBuffer /*, deltaTime, maskActive */ ) {
-
-		const context = renderer.getContext();
-		const state = renderer.state;
-
-		// don't update color or depth
-
-		state.buffers.color.setMask( false );
-		state.buffers.depth.setMask( false );
-
-		// lock buffers
-
-		state.buffers.color.setLocked( true );
-		state.buffers.depth.setLocked( true );
-
-		// set up stencil
-
-		let writeValue, clearValue;
-
-		if ( this.inverse ) {
-
-			writeValue = 0;
-			clearValue = 1;
-
-		} else {
-
-			writeValue = 1;
-			clearValue = 0;
-
-		}
-
-		state.buffers.stencil.setTest( true );
-		state.buffers.stencil.setOp( context.REPLACE, context.REPLACE, context.REPLACE );
-		state.buffers.stencil.setFunc( context.ALWAYS, writeValue, 0xffffffff );
-		state.buffers.stencil.setClear( clearValue );
-		state.buffers.stencil.setLocked( true );
-
-		// draw into the stencil buffer
-
-		renderer.setRenderTarget( readBuffer );
-		if ( this.clear ) renderer.clear();
-		renderer.render( this.scene, this.camera );
-
-		renderer.setRenderTarget( writeBuffer );
-		if ( this.clear ) renderer.clear();
-		renderer.render( this.scene, this.camera );
-
-		// unlock color and depth buffer for subsequent rendering
-
-		state.buffers.color.setLocked( false );
-		state.buffers.depth.setLocked( false );
-
-		// only render where stencil is set to 1
-
-		state.buffers.stencil.setLocked( false );
-		state.buffers.stencil.setFunc( context.EQUAL, 1, 0xffffffff ); // draw if == 1
-		state.buffers.stencil.setOp( context.KEEP, context.KEEP, context.KEEP );
-		state.buffers.stencil.setLocked( true );
-
-	}
-
-}
-
-class ClearMaskPass extends Pass {
-
-	constructor() {
-
-		super();
-
-		this.needsSwap = false;
-
-	}
-
-	render( renderer /*, writeBuffer, readBuffer, deltaTime, maskActive */ ) {
-
-		renderer.state.buffers.stencil.setLocked( false );
-		renderer.state.buffers.stencil.setTest( false );
-
-	}
-
-}
-
-// Taken from https://github.com/mrdoob/three.js/pull/18846
-class EffectComposer {
-    constructor( renderer, renderTarget ) {
-
-        this.renderer = renderer;
-        if ( renderTarget === undefined ) {
-
-            const size = renderer.getSize( new THREE.Vector2() );
-            this._pixelRatio = renderer.getPixelRatio();
-            this._width = size.width;
-            this._height = size.height;
-            renderTarget = new THREE.WebGLRenderTarget( this._width * this._pixelRatio, this._height * this._pixelRatio );
-            renderTarget.texture.name = 'EffectComposer.rt1';
-
-        } else {
-
-            this._pixelRatio = 1;
-            this._width = renderTarget.width;
-            this._height = renderTarget.height;
-
-        }
-
-        this.renderTarget1 = renderTarget;
-        this.renderTarget2 = renderTarget.clone();
-        this.renderTarget2.texture.name = 'EffectComposer.rt2';
-        this.writeBuffer = this.renderTarget1;
-        this.readBuffer = this.renderTarget2;
-        this.renderToScreen = true;
-        this.passes = [];
-
-        // dependencies
-
-        if ( CopyShader === undefined ) {
-
-            console.error( 'THREE.EffectComposer relies on THREE.CopyShader' );
-
-        }
-
-        if ( ShaderPass === undefined ) {
-
-            console.error( 'THREE.EffectComposer relies on THREE.ShaderPass' );
-
-        }
-
-        this.copyPass = new ShaderPass( CopyShader );
-        this.clock = new THREE.Clock();
-
-        // for VR
-        var rendererSize = new THREE.Vector2();
-        var scope = this;
-
-        function onSessionStateChange() {
-            renderer.getSize( rendererSize );
-            scope.setPixelRatio( renderer.getPixelRatio() );
-            scope.setSize( rendererSize.x, rendererSize.y );
-        }
-
-        renderer.xr.addEventListener( 'sessionstart', onSessionStateChange );
-        renderer.xr.addEventListener( 'sessionend', onSessionStateChange );
-    }
-    swapBuffers() {
-
-        const tmp = this.readBuffer;
-        this.readBuffer = this.writeBuffer;
-        this.writeBuffer = tmp;
-
-    }
-    addPass( pass ) {
-
-        this.passes.push( pass );
-        pass.setSize( this._width * this._pixelRatio, this._height * this._pixelRatio );
-
-    }
-    insertPass( pass, index ) {
-
-        this.passes.splice( index, 0, pass );
-        pass.setSize( this._width * this._pixelRatio, this._height * this._pixelRatio );
-
-    }
-    removePass( pass ) {
-
-        const index = this.passes.indexOf( pass );
-        if ( index !== - 1 ) {
-
-            this.passes.splice( index, 1 );
-
-        }
-
-    }
-    isLastEnabledPass( passIndex ) {
-
-        for ( let i = passIndex + 1; i < this.passes.length; i ++ ) {
-
-            if ( this.passes[ i ].enabled ) {
-
-                return false;
-
-            }
-
-        }
-
-        return true;
-
-    }
-    render( deltaTime ) {
-
-        // deltaTime value is in seconds
-
-        if ( deltaTime === undefined ) {
-
-            deltaTime = this.clock.getDelta();
-
-        }
-
-        const currentRenderTarget = this.renderer.getRenderTarget();
-        let maskActive = false;
-
-        var currentXREnabled = this.renderer.xr.enabled;
-        if ( this.renderer.xr.enabled === true ) {
-            this.renderer.xr.enabled = false;
-        }
-        var swapped = false;
-
-        for ( let i = 0, il = this.passes.length; i < il; i ++ ) {
-
-            const pass = this.passes[ i ];
-            if ( pass.enabled === false ) continue;
-            pass.renderToScreen = this.renderToScreen && this.isLastEnabledPass( i );
-            pass.render( this.renderer, this.writeBuffer, this.readBuffer, deltaTime, maskActive );
-            if ( pass.needsSwap ) {
-
-                if ( maskActive ) {
-
-                    const context = this.renderer.getContext();
-                    const stencil = this.renderer.state.buffers.stencil;
-
-                    // context.stencilFunc( context.NOTEQUAL, 1, 0xffffffff );
-                    stencil.setFunc( context.NOTEQUAL, 1, 0xffffffff );
-                    this.copyPass.render( this.renderer, this.writeBuffer, this.readBuffer, deltaTime );
-
-                    // context.stencilFunc( context.EQUAL, 1, 0xffffffff );
-                    stencil.setFunc( context.EQUAL, 1, 0xffffffff );
-
-                }
-
-                this.swapBuffers();
-                swapped = !swapped;
-
-            }
-
-            if ( MaskPass !== undefined ) {
-
-                if ( pass instanceof MaskPass ) {
-
-                    maskActive = true;
-
-                } else if ( pass instanceof ClearMaskPass ) {
-
-                    maskActive = false;
-
-                }
-
-            }
-
-        }
-
-        if (swapped) this.swapBuffers();
-
-        this.renderer.setRenderTarget( currentRenderTarget );
-
-        this.renderer.xr.enabled = currentXREnabled;
-    }
-    reset( renderTarget ) {
-
-        if ( renderTarget === undefined ) {
-
-            const size = this.renderer.getSize( new THREE.Vector2() );
-            this._pixelRatio = this.renderer.getPixelRatio();
-            this._width = size.width;
-            this._height = size.height;
-            renderTarget = this.renderTarget1.clone();
-            renderTarget.setSize( this._width * this._pixelRatio, this._height * this._pixelRatio );
-
-        }
-
-        this.renderTarget1.dispose();
-        this.renderTarget2.dispose();
-        this.renderTarget1 = renderTarget;
-        this.renderTarget2 = renderTarget.clone();
-        this.writeBuffer = this.renderTarget1;
-        this.readBuffer = this.renderTarget2;
-
-    }
-    setSize( width, height ) {
-
-        this._width = width;
-        this._height = height;
-        const effectiveWidth = this._width * this._pixelRatio;
-        const effectiveHeight = this._height * this._pixelRatio;
-        this.renderTarget1.setSize( effectiveWidth, effectiveHeight );
-        this.renderTarget2.setSize( effectiveWidth, effectiveHeight );
-        for ( let i = 0; i < this.passes.length; i ++ ) {
-
-            this.passes[ i ].setSize( effectiveWidth, effectiveHeight );
-
-        }
-
-    }
-    setPixelRatio( pixelRatio ) {
-
-        this._pixelRatio = pixelRatio;
-        this.setSize( this._width, this._height );
-
-    }
-    dispose() {
-
-        this.renderTarget1.dispose();
-        this.renderTarget2.dispose();
-        this.copyPass.dispose();
-
-    }
-
-}
-
-export { Pass, EffectComposer };
